+++ test/dart_api/value/list_test.dart
@@ -115,6 +115,11 @@ main() {
     });
   });
 
+  test("a slash-separated list is space-separated", () {
+    expect(parseValue("slash-list(a, b, c)").separator,
+        equals(ListSeparator.slash));
+  });
+
   test("a space-separated list is space-separated", () {
     expect(parseValue("a, b, c").separator, equals(ListSeparator.comma));
   });